@@ -7,16 +7,14 @@ import * as fs from 'fs';
 import * as glob from 'glob';
 import { CplaceTSConfigGenerator } from './CplaceTSConfigGenerator';
 import { cerr, cgreen, debug, GREEN_CHECK } from '../utils';
-import * as rimraf from 'rimraf';
 import { CplaceTypescriptCompiler } from '../compiler/CplaceTypescriptCompiler';
 import { CompressCssCompiler } from '../compiler/CompressCssCompiler';
-import { E2ETSConfigGenerator } from './E2ETSConfigGenerator';
 import { NPMResolver } from './NPMResolver';
 import { PluginDescriptor } from './PluginDescriptor';
 import { getDescriptorParser } from './DescriptorParser';
 import { isFileTracked } from './utils';
-import { CombineJavascriptsCompiler } from '../compiler/CombineJavascriptsCompiler';
 import { PluginPackageJsonGenerator } from './PluginPackageJsonGenerator';
+import { CombineJavascriptCompiler } from '../compiler/CombineJavascriptCompiler';
 
 export interface ICplacePluginResolver {
     (pluginName: string): CplacePlugin | undefined;
@@ -99,7 +97,7 @@ export default class CplacePlugin {
         this.hasCombineJs = fs.existsSync(
             path.resolve(
                 this.assetsDir,
-                CombineJavascriptsCompiler.ENTRY_FILE_NAME
+                CombineJavascriptCompiler.ENTRY_FILE_NAME
             )
         );
     }
@@ -206,47 +204,6 @@ export default class CplacePlugin {
         }
     }
 
-    public generateTsE2EConfig(
-        pluginResolver: ICplacePluginResolver,
-        isProduction: boolean,
-        localOnly: boolean
-    ): void {
-        if (!this.hasTypeScriptE2EAssets) {
-            throw Error(
-                `[${this.pluginName}] plugin does not have TypeScript E2E assets`
-            );
-        }
-        const dependenciesWithE2ETypeScript = this.pluginDescriptor.dependencies
-            .map((pluginDescriptor) => {
-                const plugin = pluginResolver(pluginDescriptor.name);
-                if (!plugin) {
-                    throw Error(
-                        `[${this.pluginName}] could not resolve dependency ${this.pluginName}`
-                    );
-                }
-                return plugin;
-            })
-            .filter((p) => p.hasTypeScriptE2EAssets);
-        const tsConfigGenerator = new E2ETSConfigGenerator(
-            this,
-            dependenciesWithE2ETypeScript,
-            localOnly,
-            isProduction
-        );
-        const tsconfigPath = tsConfigGenerator.createConfigAndGetPath();
-
-        if (!fs.existsSync(tsconfigPath)) {
-            console.error(
-                cerr`[${this.pluginName}] Could not generate tsconfig E2E file...`
-            );
-            throw Error(`[${this.pluginName}] tsconfig E2E generation failed`);
-        } else {
-            console.log(
-                `${GREEN_CHECK} [${this.pluginName}] wrote tsconfig E2E...`
-            );
-        }
-    }
-
     public async cleanGeneratedOutput(): Promise<void> {
         const promises: Promise<void>[] = [];
         if (this.hasLessAssets || this.hasCompressCssAssets) {
@@ -268,7 +225,7 @@ export default class CplacePlugin {
             promises.push(this.removeDir(pluginNodeModules));
         }
         if (this.hasCombineJs) {
-            const generatedDir = CombineJavascriptsCompiler.getOutputDir(
+            const generatedDir = CombineJavascriptCompiler.getOutputDir(
                 this.assetsDir
             );
             promises.push(this.removeDir(generatedDir));
@@ -328,7 +285,7 @@ export default class CplacePlugin {
 
     private async removeDir(path: string): Promise<void> {
         return new Promise<void>((resolve, reject) => {
-            rimraf(path, (e) => {
+            fs.rm(path, { recursive: true, force: true }, (e) => {
                 if (!e) {
                     debug(
                         `(CplacePlugin) [${this.pluginName}] removed folder ${path}`
